Remove alert inclusion from signup page
The alert component inclusion was removed from the signup page to avoid confusion among users. Previously, users who tried to signup with an existing email were advised to log in. However, this led to an unclear user guidance process. Now, with the removal of this alert, users will receive a more straightforward prompt to use a different email for signing up.