Change user field name and admin account details
The 'username' field in the Users table was changed to 'fullname' to accurately reflect the data it holds. The admin account details in the constant.php file were also updated to match these changes. This refactoring was done to improve the clarity and consistency of the codebase and the user's information representation.