Refactor employee table and form UI
The employee table and form UI have been refactored for simplifying user interaction and improving overall experience. The 'EmployeeID', 'FirstName', and 'LastName' columns were merged into a singular 'Employee Name' column to more concisely present data. The presentation layer of the application has also been modified to now show the form on the side of the table for a more streamlined look and a better UX, particularly on horizontal displays. This refactoring improves the usability of the application without affecting its functionality.